+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## [Unreleased][unreleased]
+
+## 1.1.3 - 2015-04-20
 ### Removed
 - click from requireds
 
@@ -48,7 +50,8 @@
 - cloning.py and process.py as first working concept
 - initial commit with necessary files
 
-[unreleased](https://github.com/Friz-zy/pyspaces/compare/v1.1.2...HEAD)
+[unreleased](https://github.com/Friz-zy/pyspaces/compare/v1.1.3...HEAD)
+[1.1.2-1.1.3](https://github.com/Friz-zy/pyspaces/compare/v1.1.2...v1.1.3)
 [1.1.1-1.1.2](https://github.com/Friz-zy/pyspaces/compare/v1.1.1...v1.1.2)
 [1.1-1.1.1](https://github.com/Friz-zy/pyspaces/compare/v1.1...v1.1.1)
 [1.0.5-1.1](https://github.com/Friz-zy/pyspaces/compare/v1.0.5...v1.1)